--- jpeg/imagedata.js
@@ -1,9 +1,0 @@
-function readJpegGenericSegment(offset, buffer) {
-	var view = new DataView(buffer);
-	var array = new Uint8Array(buffer);
-
-	var length = view.getUint16(offset + 2);
-	var end = length + offset + 2;
-
-	return Array.prototype.slice.call(array).slice(offset, end);
-}
