--- src/app1/index.js
@@ -1,3 +1,0 @@
-export readId from './id'
-export readAdobeXML from './adobe'
-export readLength from './length'
