+++ src/dict/iptc/index.js
@@ -1,0 +1,2 @@
+export envelope from './envelope'
+export application2 from './application2'
