+++ src/dict/tiff/index.js
@@ -1,0 +1,4 @@
+export gps from './gps'
+export image from './image'
+export iop from './iop'
+export photo from './photo'
