--- src/dictionary/index.js
@@ -1,21 +1,0 @@
-import application2 from './IPTC.APPLICATION2'
-import envelope from './IPTC.ENVELOPE'
-import image from './TIFF.IMAGE'
-import iop from './TIFF.IOP'
-import photo from './TIFF.PHOTO'
-
-
-
-export jpeg from './JPEG'
-
-export const iptc = {
-  application2,
-  envelope,
-}
-
-
-export const tiff = {
-  photo,
-  image,
-  iop,
-}
