+++ src/index.js
@@ -1,6 +1,13 @@
-export MicrosoftPadding from './MicrosoftPadding'
-export markers from './markers'
-export jfif from './jfif'
-export iptc from './iptc'
-export generic from './generic'
-export comment from './comment'
+export readAdobe from './adobe'
+export readComment from './comment'
+export readIPTC from './iptc'
+export readJFIF from './jfif'
+export readMarkers from './markers'
+export readMicrosoftPadding from './microsoftPadding'
+export readSegment from './generic'
+export exif from './exif'
+
+export util from './util'
+export app1 from './app1'
+
+export dict from './dict'
