+++ src/util/index.js
@@ -1,0 +1,4 @@
+export charsToUTF16String from './charsToUTF16String'
+export lettersToCharCodes from './lettersToCharCodes'
+export lettersToString from './lettersToString'
+export charsToString from './charsToString'
