+++ src/util/lettersToString.js
@@ -1,8 +1,8 @@
 export default function lettersToString(letters) {
     //combine characters to a string ["E","x","i","f"] -> "Exif"
-
-    let output = ''
-    letters.forEach(character => output += character)
-    return output
-    
-}
+    var output = "";
+    letters.forEach(function(character) {
+        output += character;
+    });
+    return output;
+  }
