+++ webpack.config.js
@@ -9,7 +9,7 @@ module.exports = {
 
   output: {
       path: path.join(__dirname, './dist'),
-      filename: '[name].bundle.js',
+      filename: 'index.js',
       publicPath: '/'
   },
 
@@ -52,11 +52,7 @@ module.exports = {
     new webpack.optimize.UglifyJsPlugin({
       beautify: false,
       mangle: {
-        screw_ie8: true,
         keep_fnames: true
-      },
-      compress: {
-        screw_ie8: true
       },
       comments: false
     }),
